add pathFilter parameter to Pack
This pathFilter is necessary to create package for Debian or OpenEmbedded:
the tar files inside dpkg or ipkg package must have path like "./usr/bin/foo",
i.e. with a leading './'.

pathFilter expects a function which is run on each file or directory stored
in the archive. This takes path as a parameter and is expected to return
a modified path. For instance:

fstream
	.Reader({ path: path.join(this.tempDir,  'usr'), type: "Directory" })
	.pipe(tar.Pack({
		noProprietary: true,
		pathFilter: function(p) {return './' + p ;}
	}))
	.pipe(zlib.createGzip())
	.pipe(fstream.Writer("data.tar.gz"));