queue fs.chown and fs.utimes so that they happen prior to close
This makes some benchmarks slightly slower, but prevents issues where a
chown or utimes is not happening until after the next file has started
being written.  That can be a problem, because the subsequent file in an
archive may in fact be trying to overwrite the prior file, creating a
race condition where the resulting mode/mtime is non-deterministic.