Unpack: do not emit 'end'
Since the Unpack class is a writable stream, but not a readable stream,
it doesn't make sense for it to emit an 'end' event.  What we're really
after is a 'close' event which means 'done writing all the things'.

However, since the Parse class already used 'end' in many cases, and
this event name was being consumed (and is, in general, safe if one is
merely parsing, and not doing asynchronous unpacking as well), it's
being left in place.

Many stream utilities call the stream 'done' at the first occurrence of
any end-of-stream style thing: finish, end, done, close, etc.