Add uid/gid options to extract and unpack
Enables forcibly setting the ownership of all extracted items,
regardless of the uid/gid of the process or the value in the archive.

This also avoids calling chown on files and directories if they are
already going to be owned by the proper uid/gid (that is, if the
specified uid/gid matches the process uid and gid).

Note that it does _not_ set the ownership of pre-existing folders that
are already in place when the extract is performed.  It only guarantees
that files and directories created by the unpack will have their
ownership set as specified.  Thus it provides a reliable gaurantee of
file ownership only if extracting into a previously empty directory.

Fix #133