use fs-minipass streams where appropriate
Because fs-minipass streams can be sync or async, and are capable of
doing a minimum number of syscalls and no artificial async deferrals,
they're a good fit for use in this library.

This allows for a lot of repetitive and error-prone fs code to be
eliminiated, replaced with a much cleaner stream abstraction.