unpack: on futimes/fchown fail, use utimes/chown
This avoids un-helpful warnings on AIX, and makes a better attempt at
satisfying the archive unpack contract, albeit at the cost of an extra
syscall per file on platforms that do not implement fchown and/or
futimes.

Also, with this change, uid/gid will be set in sync unpack scenarios
even if fchown fails, which removes a point of behavior divergence
between the sync and async unpack algorithms.