Apply extended headers before inferring directory type
Fix #187

Provide the extended header and global extended header to Header
creation in entry parsing.

This avoids a corner case where a NextEntryHasLongPath meta entry will
be overridden if the first bit of the path ends in a slash, such that
the Header is inferred to be a Directory type, and thus of zero length.