basic path reservation system
This is going to be used to avoid harmful race conditions in parallel
extraction operations.

For example, a Link to a/b requires that the File at a/b is present.
But, if they are run in parallel, then the File might not be written
before the Link tries to create, and it'll fail with ENOENT.

This gets worse if there's a Directory at a/b/c that unlinks the File so
that it can create a dir.