fix(unpack): fix hang on large file on open() fail
The fs-minipass WriteStream always returns false from a write() call if
it did not successfully open the file descriptor.  This makes the
Parse/Unpack stream get backed up if the file was large enough that it
did not already move on to the next entry in the archive.

Fix: https://github.com/npm/node-tar/issues/264