fix: perf regression on hot string munging path
This doesn't change any functionality, but it optimizes a few extremely
hot code paths for the input typically encountered during a large npm
project installation.

`String.normalize()` is cached, and trailing-slash removal is done with
a single `String.slice()`, rather than multiple slices and
`String.length` comparisons.

It is extremely rare that any code path is ever hot enough for this kind
of thing to be relevant enough to justify this sort of
microoptimization, but these two issues resulted in a 25-50% install
time increase in some cases, which is fairly dramatic.

Fix: https://github.com/npm/cli/issues/3676

PR-URL: https://github.com/npm/node-tar/pull/286
Credit: @isaacs
Close: #286
Reviewed-by: @wraithgar