changelog v7, flip default chmod behavior
The `noChmod` option is deprecated, and replaced with the `chmod`
option. The default is now to *not* explicitly set the mode of created
filesystem entries.

This is a deviation from `tar(1)` implementations, but prevents the need
to call `process.umask()`, which is deprecated and not thread safe.

To still set modes explicitly, while avoiding the need to call
`process.umask()`, a `processUmask` numeric option may now be provided
along with `chmod: true`.