DRY all commands, correct overload signature types
This moves the repetitive overload definitions into a single type created
by a makeCommand method. This also means that all of the repetitive setup
and validation can be moved to one place as well.

Corrects a bug where TS thought that a `Promise<void>` would be
returned, which actually it was an instance of Unpack (or whatever).

The type calculus is rather baroque. The next major will replace the
overloaded function signatures with separately named functions, so
there'll be `createFile`, `createSync`, `createFileSync`, etc.