remove dirCache from mkdir, unpack
This reduces an unnecessarily complicated and expensive combinatorially
explosive check with `O(n^2)` complexity which gets excessively slow in
cases where lots of files are being added to a large archive.

It turns out that the cost of the stats that were saved by this cache
were not worth the cost of keeping it up to date. A better data
structure *could* potentially preserve the value while avoiding the
combinatorial explosion required to keep it up to date, but it'd be
quite a lot more complicated. This can be explored if it turns out that
removing the dirCache imposes performance issues in other cases, but
this seems unlikely based on current available data.

Fix: #441