fix: consistent TOCTOU behavior in sync t.list
This brings the TOCTOU handling behavior in t.list synchronous behavior
to be the same as the async behavior, and other high-level methods.

That is, if there is a write to a file after testing its size, then the
*smaller* data size will be used, rather than potentially trying to read
too much data into memory all at once.

PR-URL: https://github.com/isaacs/node-tar/pull/444
Credit: @fabasoad
Close: #444
Reviewed-by: @isaacs